+++ shopify/packages/ajax-cart-api/src/effect/logger/LoggerUtils.ts
@@ -1,7 +1,7 @@
 import { Logging } from "@brytdesigns/shopify-utils/effect";
 
 export const withNamespacedLogSpan = Logging.makeNamespacedLogSpan(
-  "[@brytdesigns/shopify-hybrid-cart-client]"
+  "[@brytdesigns/shopify-ajax-cart-client]",
 );
 
 export const filterLevelOrNever = Logging.filterLevelOrNever;