+++ shopify/packages/ajax-cart-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brytdesigns/shopify-ajax-cart-api
 
+## 1.0.2
+
+### Patch Changes
+
+- Minor bug fixes
+
 ## 1.0.1
 
 ### Patch Changes