+++ shopify/packages/ajax-cart-components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @brytdesigns/shopify-ajax-cart-components
 
+## 1.0.5
+
+### Patch Changes
+
+- Minor bug fixes
+  - @brytdesigns/shopify-ajax-cart@1.0.3
+
+## 1.0.4
+
+### Patch Changes
+
+- Minor bug fix with cart-data-injection web component
+
 ## 1.0.3
 
 ### Patch Changes